Work on fixing docstrings and splitting modules.
