+++ aslprep/__init__.py
@@ -1,6 +1,6 @@
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
-"""Top-module metadata."""
+"""ASLPrep: A Robust Preprocessing Pipeline for ASL Data"""
 
 from aslprep.__about__ import __copyright__, __credits__, __packagename__, __version__
