+++ aslprep/interfaces/confounds.py
@@ -1,18 +1,6 @@
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
-"""
-Handling confounds.
-
-    .. testsetup::
-
-    >>> import os
-    >>> import pandas as pd
-
-"""
-
-import os
-import re
-
+"""Interfaces for calculating and collecting confounds."""
 import pandas as pd
 from nipype import logging
 from nipype.interfaces.base import (
@@ -24,6 +12,9 @@ from nipype.interfaces.base import (
     traits,
 )
 from nipype.utils.filemanip import fname_presuffix
+
+from aslprep.niworkflows.viz.plots import ASLPlot
+from aslprep.utils.confounds import _gather_confounds
 
 LOGGER = logging.getLogger("nipype.interface")
 
@@ -43,37 +34,8 @@ class _GatherConfoundsOutputSpec(TraitedSpec):
 
 
 class GatherConfounds(SimpleInterface):
-    r"""
-    Combine various sources of confounds in one TSV file
+    """Combine various sources of confounds in one TSV file."""
 
-    .. testsetup::
-
-    >>> from tempfile import TemporaryDirectory
-    >>> tmpdir = TemporaryDirectory()
-    >>> os.chdir(tmpdir.name)
-
-    .. doctest::
-
-    >>> pd.DataFrame({'a': [0.1]}).to_csv('signals.tsv', index=False, na_rep='n/a')
-    >>> pd.DataFrame({'b': [0.2]}).to_csv('dvars.tsv', index=False, na_rep='n/a')
-
-    >>> gather = GatherConfounds()
-    >>> gather.inputs.signals = 'signals.tsv'
-    >>> gather.inputs.dvars = 'dvars.tsv'
-    >>> res = gather.run()
-    >>> res.outputs.confounds_list
-    ['Global signals', 'DVARS']
-
-    >>> pd.read_csv(res.outputs.confounds_file, sep='\s+', index_col=None,
-    ...             engine='python')  # doctest: +NORMALIZE_WHITESPACE
-         a    b
-    0  0.1  0.2
-
-    .. testcleanup::
-
-    >>> tmpdir.cleanup()
-
-    """
     input_spec = _GatherConfoundsInputSpec
     output_spec = _GatherConfoundsOutputSpec
 
@@ -90,85 +52,6 @@ class GatherConfounds(SimpleInterface):
         self._results["confounds_file"] = combined_out
         self._results["confounds_list"] = confounds_list
         return runtime
-
-
-def _gather_confounds(
-    signals=None, dvars=None, std_dvars=None, fdisp=None, rmsd=None, motion=None, newpath=None
-):
-    r"""
-    Load confounds from the filenames, concatenate together horizontally
-    and save new file.
-
-    >>> from tempfile import TemporaryDirectory
-    >>> tmpdir = TemporaryDirectory()
-    >>> os.chdir(tmpdir.name)
-    >>> pd.DataFrame({'Global Signal': [0.1]}).to_csv('signals.tsv', index=False, na_rep='n/a')
-    >>> pd.DataFrame({'stdDVARS': [0.2]}).to_csv('dvars.tsv', index=False, na_rep='n/a')
-    >>> out_file, confound_list = _gather_confounds('signals.tsv', 'dvars.tsv')
-    >>> confound_list
-    ['Global signals', 'DVARS']
-
-    >>> pd.read_csv(out_file, sep='\s+', index_col=None,
-    ...             engine='python')  # doctest: +NORMALIZE_WHITESPACE
-       global_signal  std_dvars
-    0            0.1        0.2
-    >>> tmpdir.cleanup()
-
-
-    """
-
-    def less_breakable(a_string):
-        """hardens the string to different envs (i.e., case insensitive, no whitespace, '#'"""
-        return "".join(a_string.split()).strip("#")
-
-    # Taken from https://stackoverflow.com/questions/1175208/
-    # If we end up using it more than just here, probably worth pulling in a well-tested package
-    def camel_to_snake(name):
-        s1 = re.sub("(.)([A-Z][a-z]+)", r"\1_\2", name)
-        return re.sub("([a-z0-9])([A-Z])", r"\1_\2", s1).lower()
-
-    def _adjust_indices(left_df, right_df):
-        # This forces missing values to appear at the beggining of the DataFrame
-        # instead of the end
-        index_diff = len(left_df.index) - len(right_df.index)
-        if index_diff > 0:
-            right_df.index = range(index_diff, len(right_df.index) + index_diff)
-        elif index_diff < 0:
-            left_df.index = range(-index_diff, len(left_df.index) - index_diff)
-
-    all_files = []
-    confounds_list = []
-    for confound, name in (
-        (signals, "Global signals"),
-        (std_dvars, "Standardized DVARS"),
-        (dvars, "DVARS"),
-        (fdisp, "Framewise displacement"),
-        (rmsd, "RMSD"),
-        (motion, "Motion parameters"),
-    ):
-        if confound is not None and isdefined(confound):
-            confounds_list.append(name)
-            if os.path.exists(confound) and os.stat(confound).st_size > 0:
-                all_files.append(confound)
-
-    confounds_data = pd.DataFrame()
-    for file_name in all_files:  # assumes they all have headings already
-        new = pd.read_csv(file_name, sep="\t")
-        for column_name in new.columns:
-            new.rename(
-                columns={column_name: camel_to_snake(less_breakable(column_name))}, inplace=True
-            )
-
-        _adjust_indices(confounds_data, new)
-        confounds_data = pd.concat((confounds_data, new), axis=1)
-
-    if newpath is None:
-        newpath = os.getcwd()
-
-    combined_out = os.path.join(newpath, "confounds.tsv")
-    confounds_data.to_csv(combined_out, sep="\t", index=False, na_rep="n/a")
-
-    return combined_out, confounds_list
 
 
 class _ASLSummaryInputSpec(BaseInterfaceInputSpec):
@@ -193,16 +76,15 @@ class _ASLSummaryOutputSpec(TraitedSpec):
 
 
 class ASLSummary(SimpleInterface):
-    """
-    Copy the x-form matrices from `hdr_file` to `out_file`.
+    """Copy the x-form matrices from `hdr_file` to `out_file`.
+
+    Clearly that's wrong.
     """
 
     input_spec = _ASLSummaryInputSpec
     output_spec = _ASLSummaryOutputSpec
 
     def _run_interface(self, runtime):
-        from aslprep.niworkflows.viz.plots import ASLPlot
-
         self._results["out_file"] = fname_presuffix(
             self.inputs.in_func, suffix="_aslplot.svg", use_ext=False, newpath=runtime.cwd
         )
