+++ aslprep/interfaces/multiecho.py
@@ -1,20 +1,6 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
-"""
-Multi-echo EPI
-~~~~~~~~~~~~~~
-directly from FMRIPREP
-For using multi-echo EPI data.
-
-Change directory to provide relative paths for doctests
->>> import os
->>> filepath = os.path.dirname( os.path.realpath( __file__ ) )
->>> datadir = os.path.realpath(os.path.join(filepath, '../data/'))
->>> os.chdir(datadir)
-
-"""
+"""Interfaces for working with multi-echo data."""
 import os
 
 from nipype import logging
@@ -65,9 +51,7 @@ class _T2SMapOutputSpec(TraitedSpec):
 
 
 class T2SMap(CommandLine):
-    """
-    Runs the tedana T2* workflow to generate an adaptive T2* map and create
-    an optimally combined ME-EPI time series.
+    """Run the tedana T2* workflow to generate a T2* map and create a combined time series.
 
     Example
     =======
